python/requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
tensorflow>=2.16.0
opencv-python>=4.10.0
mediapipe>=0.10.0
matplotlib>=3.9.0
seaborn>=0.13.0
flask>=3.0.0
flask-cors>=4.0.0
kagglehub>=0.3.0
reportlab>=4.2.0
pillow>=10.0.0
